// my-app/src/types.ts
export enum Label {
    personal = "personal",
    study = "study",
    work = "work",
    other = "other",
 }
 
 export type Note = {
    id: number;
    title: string;
    content: string;
    label: Label;
    favorite: boolean;
    liked: false;
 };
 
 export type GroceryItem = { name: string; isPurchased: boolean };
 